refactor | Main update

+++ Microservico_B/src/main/java/com/example/microservico_b/controller/PostController.java
@@ -97,7 +97,7 @@ public class PostController implements Serializable {
 
     )
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deletePost(@PathVariable Long id){
+    public ResponseEntity<?> deletePost(@PathVariable int id){
         try {
             postService.deletePost(id);
             return ResponseEntity.ok("Deleted with Success");
